requirements.txt
google-generativeai>=0.5.0
faiss-cpu
python-dotenv
PyPDF2
numpy
